remove user_socket file from assets directory and transfer its contents to app.js
